requirements.txt
azure-ai-inference==1.0.0b6
azure-ai-projects==1.0.0b2
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.6.0b7

numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
textstat==0.7.4
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2

tqdm==4.66.6
requests==2.32.3
regex==2024.9.11
typing_extensions==4.12.2
pydantic==2.10.6
python-dotenv==1.0.1
nltk==3.9.1

ipython==8.29.0
ipykernel==6.29.5

openai==1.60.0
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
